adding autocomplete support to bunsen
