+++ addon/index.js
@@ -10,6 +10,7 @@ export {default as Form} from './components/form'
 export {default as InputWrapper} from './components/input-wrapper'
 export {default as Section} from './components/section'
 export {default as ValidationResult} from './components/validation-result'
+import {default as Autocomplete} from './components/inputs/autocomplete'
 import {default as Boolean} from './components/inputs/boolean'
 import {default as ButtonGroup} from './components/inputs/button-group'
 import {default as CheckboxArray} from './components/inputs/checkbox-array'
@@ -30,6 +31,7 @@ import {default as Url} from './components/inputs/url'
 import {default as When} from './components/inputs/when'
 
 export const Inputs = {
+  Autocomplete,
   Boolean,
   ButtonGroup,
   CheckboxArray,
