+++ app/components/frost-bunsen-input-autocomplete.js
@@ -1,0 +1,1 @@
+export {default} from 'ember-frost-bunsen/components/inputs/autocomplete'
