+++ tests/dummy/mirage/data/models/autocomplete.js
@@ -1,0 +1,15 @@
+export default {
+  type: 'object',
+  properties: {
+    queryExampleWithCountries: {
+      type: 'string',
+      modelType: 'country',
+      valueAttribute: 'id',
+      labelAttribute: 'name',
+      query: {
+        p: 'name:$filter'
+      }
+    }
+  },
+  required: ['queryExampleWithCountries']
+}
